Moves status cache serde-for-snapshots code into own module
